+++ kafka/cflt_client.go
@@ -14,13 +14,13 @@ import (
 
 // Kafka Message Provider implementation that uses the standard Confluent golang client
 
-func NewMessageProviderFactory(topicName string, props map[string]string, groupID string) MessageProviderFactory {
-	return &ConfluentMessageProviderFactory{
-		topicName: topicName,
-		props:     props,
-		groupID:   groupID,
-	}
-}
+// func NewMessageProviderFactory(topicName string, props map[string]string, groupID string) MessageProviderFactory {
+// 	return &ConfluentMessageProviderFactory{
+// 		topicName: topicName,
+// 		props:     props,
+// 		groupID:   groupID,
+// 	}
+// }
 
 type ConfluentMessageProviderFactory struct {
 	topicName string
